@@ -18,7 +18,7 @@
 			use:melt={$overlay}
 			class="fixed inset-0 z-50 bg-black/50"
 			transition:fade={{ duration: 500 }}
-		/>
+		></div>
 		<div
 			class="fixed left-1/2 top-1/2 z-50 max-h-[85vh] w-[90vw] max-w-[450px] -translate-x-1/2 -translate-y-1/2 rounded-xl bg-white p-6 shadow-lg"
 			use:melt={$content}
